EPMRPP-37189: Code Review fixes - 1
