+++ app/src/controllers/testItem/index.js
@@ -25,4 +25,5 @@ export {
   queryParametersSelector,
   itemsSelector,
   testCaseNameLinkSelector,
+  paginationSelector,
 } from './selectors';
